+++ NEWS.md
@@ -1,6 +1,8 @@
 # bcgwcat 0.6.0
 * Rename to bcgwcat to reflect change in scope
   * Update docs etc. in light of this
+* Added more details to Shiny app explaining private vs. public EMS data, 
+  water quality guidelines and parameter/unit changes
 
 # rems2aquachem 0.5.1
 * Speed increases